Update plugin to work on OS X
sudo is required to successfully run the django tests on OS X. Plugin
now checks if the user is running Linux and if so will run the commands
without sudo and if they are using OS X then the commands will be issued
with sudo. The user will need to input their password the first time
this is done.