Feature: Single config file per project and nested test dirs work
Now only one config file is require for a project app names are listed
out in the config file. Nested test directories work without any extra
work on the part of the user.

Also fixed a bug where the wrong test would be ran if the current line
existed more than once if a file.